springboot-start/src/main/java/com/zhousuhang/demo/springboot/bean/Book.java
package com.zhousuhang.demo.springboot.bean;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

@Component
public class Book {
	@Value("${book.id}")
	private int id;
	@Value("${book.name}")
	private String bookName;
	@Value("${book.author}")
	private String author;
	@Value("${book.isbn}")
	private String isbn;

	public int getId() {
		return id;
	}

	public void setId(int id) {
		this.id = id;
	}

	public String getBookName() {
		return bookName;
	}

	public void setBookName(String bookName) {
		this.bookName = bookName;
	}

	public String getAuthor() {
		return author;
	}

	public void setAuthor(String author) {
		this.author = author;
	}

	public String getIsbn() {
		return isbn;
	}

	public void setIsbn(String isbn) {
		this.isbn = isbn;
	}

}
